+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/CallStackFactoryV1.java
@@ -23,9 +23,9 @@ public class CallStackFactoryV1 implements CallStackFactory<SpanEvent> {
 
     private final CallStack.Factory<SpanEvent> factory = new SpanEventFactory();
     private final int maxDepth;
-    private final short maxSequence;
+    private final int maxSequence;
 
-    public CallStackFactoryV1(int maxDepth, short maxSequence) {
+    public CallStackFactoryV1(int maxDepth, int maxSequence) {
         this.maxDepth = maxDepth;
         this.maxSequence = maxSequence;
     }
